logger now correctly creates log.txt if not found, minor tweaks

+++ server/logger.js
@@ -1,6 +1,5 @@
 //This is a really simple logging feature that logs to log.txt (server folder)
-//If you don't have a log.txt create one with this as the first line:
-//Timestamp	Type	Source	Message
+//Log is automatically created if it doesn't exist.
 
 var fs = require("fs");
 
@@ -9,15 +8,32 @@ var fs = require("fs");
 function log(color, source, message)
 {
 	console.log(color + source + ":\033[0m\t" + message);
-	fs.open("log.txt", "a", 666, function(e, id)
-	{
-		var kind = "Information";
-		if (color == '\033[93m') kind = "Warning";
-		else if (color == '\033[91m') kind = "Error"
-		fs.write(id, "\n" + prettyDate() + "\t" + kind + "\t" + source + "\t" + message, null, "utf8", function()
-		{
-			fs.close(id, function(){});
-		});
+
+	// Check if log.txt exists
+	fs.exists("log.txt", function(exists) {
+
+		if(exists) {
+			fs.open("log.txt", "a", 666, function(e, id) {
+				var kind = "Information";
+				if (color == '\033[93m') kind = "Warning";
+				else if (color == '\033[91m') kind = "Error";
+				fs.write(id, "\n" + prettyDate() + "\t" + kind + "\t" + source + "\t" + message, null, "utf8", function()
+				{
+					fs.close(id, function(){});
+				});
+			}
+		)
+		} 
+
+		else {
+			console.log("doesnt exist");
+			fs.open("log.txt", "w", 0666, function(err, fd) {
+				fs.write(fd, "Timestamp\tType\tSource\tMessage", null, "utf8", function()
+				{
+					fs.close(fd, function(){});
+				});
+			});
+		}
 	});
 }
 
@@ -30,7 +46,7 @@ function w(source, message)
 //Error function
 function e(source, message)
 {
-	log('\033[91m', source, message)
+	log('\033[91m', source, message);
 }
 
 //Information function
